requirements.txt
numpy
pandas
requests
notebook
matplotlib
seaborn
python-dotenv